Fix header

+++ src/libx68k/x68k/iocs.h
@@ -4,7 +4,7 @@
 #include <sys/cdefs.h>
 #include <sys/types.h>
 
-typedef unsigned int iocs_color_t;
+typedef uint16_t iocs_color_t;
 
 struct iocs_boxptr {
   short	x1;
